persistence/requirements.txt
kafka-python-ng==2.2.0
psycopg[binary]==3.1.18
